airline/requirements.txt
django
django-cors-headers
django-widget-tweaks
djangorestframework
pillow
psycopg2-binary